+++ spec/options_builder/data_labels_spec.rb
@@ -2,64 +2,63 @@ require 'spec_helper'
 
 RSpec.describe '#build_data_labels' do
   let(:x_sample) { nil }
-  let(:boolean_dl) {
-    {
-      data_labels: true
-    }
-  }
-  let(:hash_dl) {
-    {
-      data_labels: {
-        enabled: true,
-        drop_shadow: {
-          enabled: false,
-          top: 1,
-          left: 1,
-          blur: 1,
-          opacity: 0.45
-        }
-      }
-    }
-  }
-  let(:boolean_ob) {
-    Apexcharts::OptionsBuilder.new(x_sample, boolean_dl)
-  }
-  let(:hash_ob) {
-    Apexcharts::OptionsBuilder.new(x_sample, hash_dl)
-  }
-  let(:expected_boolean_built) {
-    {
-      dataLabels: {
-        enabled: true
-      }
-    }
-  }
-  let(:expected_hash_built) {
-    {
-      dataLabels: {
-        enabled: true,
-        dropShadow: {
-          enabled: false,
-          top: 1,
-          left: 1,
-          blur: 1,
-          opacity: 0.45
-        }
-      }
-    }
+  let(:ob) {
+    Apexcharts::OptionsBuilder.new(x_sample, options)
   }
 
   context "boolean data_labels" do
+    let(:options) {
+      {
+        data_labels: true
+      }
+    }
+    let(:expected_built) {
+      {
+        dataLabels: {
+          enabled: true
+        }
+      }
+    }
+
     it 'adds enabled key with the boolean as the value' do
-      boolean_ob.build_data_labels
-      expect(boolean_ob.built).to match(hash_including(expected_boolean_built))
+      ob.build_data_labels
+      expect(ob.built).to match(hash_including(expected_built))
     end
   end
 
   context "hash data_labels" do
+    let(:options) {
+      {
+        data_labels: {
+          enabled: true,
+          drop_shadow: {
+            enabled: false,
+            top: 1,
+            left: 1,
+            blur: 1,
+            opacity: 0.45
+          }
+        }
+      }
+    }
+    let(:expected_built) {
+      {
+        dataLabels: {
+          enabled: true,
+          dropShadow: {
+            enabled: false,
+            top: 1,
+            left: 1,
+            blur: 1,
+            opacity: 0.45
+          }
+        }
+      }
+    }
+
     it 'only camelizes the hash' do
-      hash_ob.build_data_labels
-      expect(hash_ob.built).to match(hash_including(expected_hash_built))
+      ob.build_data_labels
+      expect(ob.built).to match(hash_including(expected_built))
     end
   end
 end