+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 
 
-## [Unreleased]
+## [0.1.5] - 2019-07-20
 ### Added
 - Data formats on README
 - HTML+ERB support on README
@@ -89,7 +89,8 @@ All notable changes to this project will be documented in this file.
 - **Mixed charts**
 - **Syncing charts**
 
-[Unreleased]: https://github.com/styd/apexcharts.rb/compare/v0.1.4...HEAD
+[Unreleased]: https://github.com/styd/apexcharts.rb/compare/v0.1.5...HEAD
+[0.1.5]: https://github.com/styd/apexcharts.rb/compare/v0.1.4...v0.1.5
 [0.1.4]: https://github.com/styd/apexcharts.rb/compare/v0.1.3...v0.1.4
 [0.1.3]: https://github.com/styd/apexcharts.rb/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/styd/apexcharts.rb/compare/v0.1.1...v0.1.2